replace `ParquetDataset`
